feat: export service entrypoint
Move the entrypoint logic into its own package, so it can be imported from other projects.
